refactor(app): extract query functions to separate folder
